Hoist install_test_trace_subscriber to test_util/
