+++ janus_server/src/datastore.rs
@@ -2648,15 +2648,16 @@ mod tests {
         },
         message::{test_util::new_dummy_report, ReportShareError},
         task::{test_util::new_dummy_task, VdafInstance, PRIO3_AES128_VERIFY_KEY_LENGTH},
-        trace::test_util::install_test_trace_subscriber,
     };
-    use ::janus_test_util::{dummy_vdaf, generate_aead_key, MockClock};
     use assert_matches::assert_matches;
     use chrono::NaiveDate;
     use futures::future::try_join_all;
     use janus::{
         hpke::{self, associated_data_for_aggregate_share, HpkeApplicationInfo, Label},
         message::{Duration, ExtensionType, HpkeConfigId, Interval, Role, Time},
+    };
+    use janus_test_util::{
+        dummy_vdaf, generate_aead_key, install_test_trace_subscriber, MockClock,
     };
     use prio::{
         field::{Field128, Field64},
