+++ test_util/src/lib.rs
@@ -9,7 +9,9 @@ use prio::{
 };
 use rand::{thread_rng, Rng};
 use ring::aead::{LessSafeKey, UnboundKey, AES_128_GCM};
-use std::sync::{Arc, Mutex};
+use std::sync::{Arc, Mutex, Once};
+use tracing_log::LogTracer;
+use tracing_subscriber::{prelude::*, EnvFilter, Registry};
 
 pub mod dummy_vdaf;
 
@@ -215,4 +217,26 @@ where
             pts.push(vdaf.prepare_step(prep_state, prep_msg.clone()).unwrap());
         }
     }
+}
+
+/// Install a tracing subscriber for use in unit tests. This should be called at the beginning
+/// of any test that requires a tracing subscriber.
+pub fn install_test_trace_subscriber() {
+    static INSTALL_TRACE_SUBSCRIBER: Once = Once::new();
+    INSTALL_TRACE_SUBSCRIBER.call_once(|| {
+        let stdout_filter = EnvFilter::from_default_env();
+        let layer = tracing_subscriber::fmt::layer()
+            .with_thread_ids(true)
+            .with_level(true)
+            .with_target(true)
+            .with_file(true)
+            .with_line_number(true)
+            .pretty()
+            .with_test_writer()
+            .with_filter(stdout_filter);
+        let subscriber = Registry::default().with(layer);
+        tracing::subscriber::set_global_default(subscriber).unwrap();
+
+        LogTracer::init().unwrap();
+    })
 }
